재훈 - refactor : 문제 요구사항에 맞춰 8-1번 수정
